+++ runtime/runtime-core/common/test/aws/smithy/kotlin/runtime/content/ByteStreamFlowTest.kt
@@ -4,9 +4,6 @@
  */
 package aws.smithy.kotlin.runtime.content
 
-import aws.smithy.kotlin.runtime.io.SdkByteReadChannel
-import aws.smithy.kotlin.runtime.io.SdkSource
-import aws.smithy.kotlin.runtime.io.source
 import io.kotest.matchers.string.shouldContain
 import kotlinx.coroutines.*
 import kotlinx.coroutines.channels.Channel
@@ -14,27 +11,6 @@ import kotlinx.coroutines.flow.*
 import kotlinx.coroutines.test.runTest
 import java.lang.RuntimeException
 import kotlin.test.*
-
-fun interface ByteStreamFactory {
-    fun byteStream(input: ByteArray): ByteStream
-    companion object {
-        val BYTE_ARRAY: ByteStreamFactory = ByteStreamFactory { input -> ByteStream.fromBytes(input) }
-
-        val SDK_SOURCE: ByteStreamFactory = ByteStreamFactory { input ->
-            object : ByteStream.SourceStream() {
-                override fun readFrom(): SdkSource = input.source()
-                override val contentLength: Long = input.size.toLong()
-            }
-        }
-
-        val SDK_CHANNEL: ByteStreamFactory = ByteStreamFactory { input ->
-            object : ByteStream.ChannelStream() {
-                override fun readFrom(): SdkByteReadChannel = SdkByteReadChannel(input)
-                override val contentLength: Long = input.size.toLong()
-            }
-        }
-    }
-}
 
 class ByteStreamBufferFlowTest : ByteStreamFlowTest(ByteStreamFactory.BYTE_ARRAY)
 class ByteStreamSourceStreamFlowTest : ByteStreamFlowTest(ByteStreamFactory.SDK_SOURCE)
